# Remove all panics in HLL implementation, replaced with error returns